Separated helper classes into another file
